+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [1.28.6] - 2026-02-11
+
+### Added
+- analysis-enriched-formatters-codex-support
+
+## [1.28.5] - 2026-02-11
+
+### Added
+- analysis-enriched-formatters-codex-support
+
+## [1.28.4] - 2026-02-11
+
+### Added
+- current work
+
+## [1.28.3] - 2026-02-11
+
+### Added
+- current work
+
 ## [1.28.2] - 2026-02-11
 
 ### Added